Revert "Rename metrics.py to logfire.py and move configuration step there"
This reverts commit 9b62abbdafced7d33ce57dbe22188484261eaec0.
