+++ src/mainframe/endpoints/package.py
@@ -24,7 +24,7 @@ from mainframe.models.schemas import (
     QueuePackageResponse,
 )
 
-from mainframe.logfire import package_success, package_fails, packages_ingested
+from mainframe.metrics import package_success, package_fails, packages_ingested
 
 router = APIRouter(tags=["package"])
 logger: structlog.stdlib.BoundLogger = structlog.get_logger()
