+++ src/mainframe/metrics.py
@@ -1,7 +1,5 @@
 import logfire
 
-logfire.configure(send_to_logfire="if-token-present")
-
 package_success = logfire.metric_counter(
     name="package_success", description="Number of successful packages scanned", unit="packages"
 )
